# web/src/lib/api/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Athenaeum HTTP API",
    "description": "Session cookie (browser): sign in via POST /api/auth/login; cookies are sent automatically.\n\nHTTP Basic Auth: Authorization: Basic base64(username:password) — useful for OPDS clients and scripts.\n\nAPI key: Authorization: Bearer ath_\u003ctoken\u003e or X-API-Key: ath_\u003ctoken\u003e — create keys in Settings -\u003e API.\n\nJSON endpoints accept and return application/json unless noted.\n\nFile endpoints stream binary content with appropriate Content-Type and Accept-Ranges.\n\nMutating browser requests require the X-CSRF-Token header (not needed for Basic Auth or API keys).",
    "version": "dev"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
  "paths": {
...
    "/api/admin/server": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Server settings (metrics, proxy, CSP, CORS)",
        "description": "Auth: admin",
        "operationId": "GET__api_admin_server",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update server settings",
        "description": "Auth: admin",
        "operationId": "PUT__api_admin_server",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/tts": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Kokoro TTS sidecar settings",
        "description": "Auth: admin",
        "operationId": "GET__api_admin_tts",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Kokoro TTS settings",
        "description": "Auth: admin",
        "operationId": "PUT__api_admin_tts",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/tts/test": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Ping Kokoro TTS sidecar",
        "description": "Auth: admin",
        "operationId": "POST__api_admin_tts_test",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/tts/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Whether Kokoro TTS is enabled",
        "description": "Auth: optional",
        "operationId": "GET__api_tts_status",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tts/synthesize": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Synthesize speech via Kokoro sidecar",
        "description": "Auth: optional",
        "operationId": "POST__api_tts_synthesize",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tts/voices": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Kokoro voices",
        "description": "Auth: optional",
        "operationId": "GET__api_tts_voices",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header",
        "description": "API key (ath_…)"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Username and password (OPDS and scripts)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key as Bearer token"
      },
      "sessionCookie": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie",
        "description": "Browser session after POST /api/auth/login"
      }
    }
  },
...
}